+++ Source/DotNetWorkQueue.Tests/Messages/ReceivedMessageTests.cs
@@ -4,6 +4,7 @@ using System.Collections.ObjectModel;
 using AutoFixture;
 using AutoFixture.AutoNSubstitute;
 using AutoFixture.Xunit2;
+using DotNetWorkQueue.Logging;
 using DotNetWorkQueue.Messages;
 using NSubstitute;
 
@@ -19,7 +20,7 @@ namespace DotNetWorkQueue.Tests.Messages
         public void Create_MessageId_Equals()
         {
             var message = CreateMessage();
-            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp());
+            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp(), new NullLogger());
             Assert.Equal(test.MessageId, message.MessageId);
         }
 
@@ -27,7 +28,7 @@ namespace DotNetWorkQueue.Tests.Messages
         public void Create_Body_Equals()
         {
             var message = CreateMessage();
-            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp());
+            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp(), new NullLogger());
             Assert.Equal(test.Body, message.Body);
         }
 
@@ -35,7 +36,7 @@ namespace DotNetWorkQueue.Tests.Messages
         public void Create_CorrelationId_Equals()
         {
             var message = CreateMessage();
-            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp());
+            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp(), new NullLogger());
             Assert.Equal(test.CorrelationId, message.CorrelationId);
         }
 
@@ -47,7 +48,7 @@ namespace DotNetWorkQueue.Tests.Messages
             var headers = new Dictionary<string, object> {{ value, new UriBuilder()}};
 
             message.Headers.Returns(new ReadOnlyDictionary<string, object>(headers));
-            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp());
+            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp(), new NullLogger());
             Assert.Equal(test.Headers, message.Headers);
         }
 
@@ -56,7 +57,7 @@ namespace DotNetWorkQueue.Tests.Messages
         {
             var fixture = new Fixture().Customize(new AutoNSubstituteCustomization());
             var message = CreateMessage();
-            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp());
+            var test = new ReceivedMessage<FakeMessage>(message, new GetPreviousErrorsNoOp(), new NullLogger());
 
             var messageContextDataFactory = fixture.Create<IMessageContextDataFactory>();